fix: MTTR ช่างกับ avg-resolution การ์ด summary ให้ปัดเป็นชั่วโมงตรงกัน (เลิก double-rounding)
งานชุดเดียวกันเคยแสดง avg แก้ไข 1.1 ชม. บนการ์ด summary แต่ 1.2 ชม. ในตาราง
ผลงานช่าง เพราะ query ช่าง `ROUND(AVG(min),0)` ปัด avg นาทีเป็นจำนวนเต็ม
"ก่อน" หารเป็นชั่วโมง (68.5→69 → 1.15 → 1.2) ส่วนการ์ด summary เก็บ 1 ตำแหน่ง
(68.5 → 1.14 → 1.1 = ค่าที่ถูก).

- ReportRepository: mttr_minutes เปลี่ยนเป็น `ROUND(AVG(min),1)` เท่ากับ query summary
- ReportService: cast mttr_minutes เป็น float (ไม่ใช่ int) เก็บทศนิยม .5 ไว้
  ให้ปัดครั้งเดียวตอนแปลงเป็นชั่วโมง — pipeline เดียวกับการ์ด summary เป๊ะ
- test: เพิ่ม regression 17+120 นาที (avg 68.5) ต้องได้ 1.1 ไม่ใช่ 1.2 (35 tests pass)

+++ app/Services/ReportService.php
@@ -113,7 +113,10 @@ class ReportService
         $resolved = (int) ($row['resolved'] ?? 0);
         $ratingCount = (int) ($row['rating_count'] ?? 0);
         $avgRating = (float) ($row['avg_rating'] ?? 0);
-        $mttrMinutes = (int) ($row['mttr_minutes'] ?? 0);
+        // float (not int): keep the .5-minute precision so MTTR rounds to hours the same way the
+        // summary card's avg_resolution does — otherwise pre-truncating minutes tips the 1-decimal
+        // hours across a boundary (68.5min→69 gave 1.2h vs the summary's correct 1.1h).
+        $mttrMinutes = (float) ($row['mttr_minutes'] ?? 0);
         $laborMinutes = (int) ($row['labor_minutes'] ?? 0);
         $completionPct = $assigned > 0 ? round($resolved / $assigned * 100, 1) : null;
